+++ anyblok_pyramid/config.py
@@ -11,7 +11,7 @@ import os
 
 @Configuration.add('preload', label="Preload")
 def define_preload_option(group):
-    group.add_argument('--databases', dest='db_names', default='',
+    group.add_argument('--databases', dest='db_names', nargs="+",
                        help='List of the database allow to be load')