+++ doc/CHANGES.rst
@@ -1,6 +1,7 @@
 .. This file is a part of the AnyBlok / Pyramid project
 ..
 ..    Copyright (C) 2015 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+..    Copyright (C) 2016 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 ..
 .. This Source Code Form is subject to the terms of the Mozilla Public License,
 .. v. 2.0. If a copy of the MPL was not distributed with this file,You can
@@ -10,6 +11,9 @@
 
 CHANGELOG
 =========
+
+* [REF] Preload database, add log and check if the database exist before load
+  it
 
 0.5.2 (2016-01-15)
 ------------------